Rename to MemoryLruCache
