+++ store/src/androidTest/kotlin/com/dropbox/external/store5/MarketMultithreadedTests.kt
@@ -6,11 +6,10 @@ import com.dropbox.external.store5.fake.FakeNotes
 import com.dropbox.external.store5.fake.OkTestMarket
 import com.dropbox.external.store5.fake.api.FakeApi
 import com.dropbox.external.store5.fake.model.Note
-import com.dropbox.external.store5.impl.ShareableLruCache
+import com.dropbox.external.store5.impl.MemoryLruCache
 import com.dropbox.external.store5.impl.SomeBroadcast
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.TestScope
-import kotlinx.coroutines.test.advanceUntilIdle
 import kotlinx.coroutines.test.runTest
 import kotlinx.coroutines.withContext
 import org.junit.Before
@@ -26,7 +25,7 @@ class MarketMultithreadedTests {
     private lateinit var api: FakeApi
     private lateinit var market: Market<String>
     private lateinit var db: FakeDb
-    private lateinit var memoryLruCache: ShareableLruCache
+    private lateinit var memoryLruCache: MemoryLruCache
     private lateinit var factory: FakeFactory<String, Note, Note>
 
     @Before
