+++ store/src/commonTest/kotlin/com/dropbox/external/store5/OfflineMarketTests.kt
@@ -8,7 +8,7 @@ import com.dropbox.external.store5.fake.FakeNotes
 import com.dropbox.external.store5.fake.OkTestMarket
 import com.dropbox.external.store5.fake.api.FakeApi
 import com.dropbox.external.store5.fake.model.Note
-import com.dropbox.external.store5.impl.ShareableLruCache
+import com.dropbox.external.store5.impl.MemoryLruCache
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.async
 import kotlinx.coroutines.test.TestScope
@@ -25,7 +25,7 @@ class OfflineMarketTests {
     private lateinit var api: FakeApi
     private lateinit var market: Market<String>
     private lateinit var db: FakeDb
-    private lateinit var memoryLruCache: ShareableLruCache
+    private lateinit var memoryLruCache: MemoryLruCache
     private lateinit var factory: FakeFactory<String, Note, Note>
 
     @BeforeTest
