+++ store/src/commonTest/kotlin/com/dropbox/external/store5/fake/BadTestMarket.kt
@@ -5,12 +5,12 @@ package com.dropbox.external.store5.fake
 import com.dropbox.external.store5.ConflictResolver
 import com.dropbox.external.store5.Store
 import com.dropbox.external.store5.fake.model.Note
-import com.dropbox.external.store5.impl.ShareableLruCache
+import com.dropbox.external.store5.impl.MemoryLruCache
 import com.dropbox.external.store5.impl.ShareableMarket
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 
 internal object BadTestMarket {
-    val memoryLruCache = ShareableLruCache(10)
+    val memoryLruCache = MemoryLruCache(10)
     val db = FakeDb()
 
     private val memoryLruCacheStore = Store.by<String, Note, Note>(
